Rename command modules namespace
