+++ src/TextToTalk/TextToTalk.cs
@@ -25,10 +25,10 @@ using TextToTalk.Backends.Polly;
 using TextToTalk.Backends.System;
 using TextToTalk.Backends.Uberduck;
 using TextToTalk.Backends.Websocket;
+using TextToTalk.CommandModules;
 using TextToTalk.Events;
 using TextToTalk.GameEnums;
 using TextToTalk.Middleware;
-using TextToTalk.Modules;
 using TextToTalk.Talk;
 using TextToTalk.TextProviders;
 using TextToTalk.UI;
